+++ src/products/components/ProductUpdatePage/types.ts
@@ -1,0 +1,148 @@
+import { OutputData } from "@editorjs/editorjs";
+import { RichTextProps } from "@saleor/attributes/utils/data";
+import { AttributeInput } from "@saleor/components/Attributes";
+import { ChannelOpts } from "@saleor/components/ChannelsAvailabilityCard/types";
+import { DatagridChangeOpts } from "@saleor/components/Datagrid/useDatagridChange";
+import { MetadataFormData } from "@saleor/components/Metadata";
+import { MultiAutocompleteChoiceType } from "@saleor/components/MultiAutocompleteSelectField";
+import { SingleAutocompleteChoiceType } from "@saleor/components/SingleAutocompleteSelectField";
+import {
+  MetadataErrorFragment,
+  ProductChannelListingUpdateInput,
+  ProductFragment,
+  SearchPagesQuery,
+  SearchProductsQuery,
+  SearchWarehousesQuery,
+} from "@saleor/graphql";
+import {
+  CommonUseFormResultWithHandlers,
+  FormChange,
+  FormErrors,
+  SubmitPromise,
+} from "@saleor/hooks/useForm";
+import {
+  FormsetAtomicData,
+  FormsetChange,
+  FormsetData,
+} from "@saleor/hooks/useFormset";
+import { UseProductUpdateHandlerError } from "@saleor/products/views/ProductUpdate/handlers/useProductUpdateHandler";
+import { FetchMoreProps, RelayToFlat, ReorderEvent } from "@saleor/types";
+
+import { ProductChannelsListingDialogSubmit } from "./ProductChannelsListingsDialog";
+
+export interface ProductUpdateFormData extends MetadataFormData {
+  category: string | null;
+  changeTaxCode: boolean;
+  chargeTaxes: boolean;
+  collections: string[];
+  isAvailable: boolean;
+  name: string;
+  rating: number;
+  slug: string;
+  seoDescription: string;
+  seoTitle: string;
+  sku: string;
+  taxCode: string;
+  trackInventory: boolean;
+  isPreorder: boolean;
+  globalThreshold: string;
+  globalSoldUnits: number;
+  hasPreorderEndDate: boolean;
+  preorderEndDateTime?: string;
+  weight: string;
+}
+export interface FileAttributeInputData {
+  attributeId: string;
+  file: File;
+}
+export type FileAttributeInput = FormsetAtomicData<
+  FileAttributeInputData,
+  string[]
+>;
+
+export interface FileAttributesSubmitData {
+  fileAttributes: FileAttributeInput[];
+}
+export interface ProductUpdateData extends ProductUpdateFormData {
+  attributes: AttributeInput[];
+  channels: ProductChannelListingUpdateInput;
+  description: OutputData;
+}
+export interface ProductUpdateSubmitData extends ProductUpdateFormData {
+  attributes: AttributeInput[];
+  attributesWithNewFileValue: FormsetData<null, File>;
+  channels: ProductChannelListingUpdateInput;
+  collections: string[];
+  description: OutputData;
+  variants: DatagridChangeOpts;
+}
+
+export interface ProductUpdateHandlers
+  extends Record<
+      | "changeMetadata"
+      | "selectCategory"
+      | "selectCollection"
+      | "selectTaxRate",
+      FormChange
+    >,
+    Record<
+      "selectAttribute" | "selectAttributeMultiple",
+      FormsetChange<string>
+    > {
+  changeChannels: (id: string, data: ChannelOpts) => void;
+  selectAttributeReference: FormsetChange<string[]>;
+  selectAttributeFile: FormsetChange<File>;
+  reorderAttributeValue: FormsetChange<ReorderEvent>;
+  changeVariants: (data: DatagridChangeOpts) => void;
+  fetchReferences: (value: string) => void;
+  fetchMoreReferences: FetchMoreProps;
+  updateChannelList: ProductChannelsListingDialogSubmit;
+}
+
+export interface UseProductUpdateFormOutput
+  extends CommonUseFormResultWithHandlers<
+      ProductUpdateData,
+      ProductUpdateHandlers
+    >,
+    RichTextProps {
+  formErrors: FormErrors<ProductUpdateSubmitData>;
+}
+
+export type UseProductUpdateFormRenderProps = Omit<
+  UseProductUpdateFormOutput,
+  "richText"
+>;
+
+export interface UseProductUpdateFormOpts
+  extends Record<
+    "categories" | "collections" | "taxTypes",
+    SingleAutocompleteChoiceType[]
+  > {
+  setSelectedCategory: React.Dispatch<React.SetStateAction<string>>;
+  setSelectedCollections: React.Dispatch<
+    React.SetStateAction<MultiAutocompleteChoiceType[]>
+  >;
+  setSelectedTaxType: React.Dispatch<React.SetStateAction<string>>;
+  selectedCollections: MultiAutocompleteChoiceType[];
+  warehouses: RelayToFlat<SearchWarehousesQuery["search"]>;
+  hasVariants: boolean;
+  referencePages: RelayToFlat<SearchPagesQuery["search"]>;
+  referenceProducts: RelayToFlat<SearchProductsQuery["search"]>;
+  fetchReferencePages?: (data: string) => void;
+  fetchMoreReferencePages?: FetchMoreProps;
+  fetchReferenceProducts?: (data: string) => void;
+  fetchMoreReferenceProducts?: FetchMoreProps;
+  assignReferencesAttributeId?: string;
+  isSimpleProduct: boolean;
+}
+
+export type SubmitResult = SubmitPromise<
+  Array<UseProductUpdateHandlerError | MetadataErrorFragment>
+>;
+
+export interface ProductUpdateFormProps extends UseProductUpdateFormOpts {
+  children: (props: UseProductUpdateFormRenderProps) => React.ReactNode;
+  product: ProductFragment;
+  onSubmit: (data: ProductUpdateSubmitData) => SubmitResult;
+  disabled: boolean;
+}
