+++ src/products/components/ProductVariants/ProductVariants.tsx
@@ -18,7 +18,7 @@ import { ProductVariantListError } from "@saleor/products/views/ProductUpdate/ha
 import React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
-import { getColumnData, getData } from "./utils";
+import { getColumnData, getData, getError } from "./utils";
 
 interface ProductVariantsProps {
   channels: ChannelData[];
@@ -74,12 +74,16 @@ export const ProductVariants: React.FC<ProductVariantsProps> = ({
       getData({
         availableColumns: columns,
         column,
-        errors,
         row,
         channels,
         variants,
         ...opts,
       }),
+    [columns, variants],
+  );
+
+  const getCellError = React.useCallback(
+    (item: Item) => getError(item, errors, columns, variants),
     [columns, variants, errors],
   );
 
@@ -92,6 +96,7 @@ export const ProductVariants: React.FC<ProductVariantsProps> = ({
       })}
       availableColumns={columns}
       getCellContent={getCellContent}
+      getCellError={getCellError}
       menuItems={index => [
         {
           label: "Edit Variant",
